Add errors handling for import data activity

+++ src/com/nilhcem/frcndict/services/ImportDataService.java
@@ -11,6 +11,8 @@ import android.util.Log;
 
 import com.nilhcem.frcndict.ApplicationController;
 import com.nilhcem.frcndict.ImportDataActivity;
+import com.nilhcem.frcndict.R;
+import com.nilhcem.frcndict.database.DatabaseHelper;
 import com.nilhcem.frcndict.utils.FileReader;
 import com.nilhcem.frcndict.utils.HttpDownloader;
 import com.nilhcem.frcndict.utils.Md5;
@@ -22,14 +24,19 @@ public final class ImportDataService {
 	private UnzipAsync unzipTask = new UnzipAsync();
 
 	public void startDownload() {
-		downloadTask.execute("");
+		downloadTask.execute();
 	}
 
 	public void setActivity(ImportDataActivity activity) {
 		this.activity = new WeakReference<ImportDataActivity>(activity);
 	}
 
-	private class DownloadFileAsync extends AsyncTask<String, Integer, String> implements Observer {
+	public void resetTasks() {
+		downloadTask = new DownloadFileAsync();
+		unzipTask = new UnzipAsync();
+	}
+
+	private class DownloadFileAsync extends AsyncTask<Void, Integer, Integer> implements Observer {
 		private static final String DICT_URL = "http://192.168.1.2/cfdict/dictionary.zip";
 		private static final String MD5_URL = "http://192.168.1.2/cfdict/md5sum";
 		private static final String TEMP_ZIP_FILE = "download.tmp";
@@ -51,30 +58,36 @@ public final class ImportDataService {
 		}
 
 		@Override
-		protected String doInBackground(String... params) {
-			HttpDownloader downloader = new HttpDownloader(DICT_URL, zipFile);
-			downloader.addObserver(this);
-			downloader.start();
-			checkMd5();
-			return null;
+		protected Integer doInBackground(Void... unused) {
+			Integer errorCode = null;
+			try {
+				HttpDownloader downloader = new HttpDownloader(DICT_URL, zipFile);
+				downloader.addObserver(this);
+				downloader.start();
+				errorCode = checkMd5();
+			} catch (Exception e) {
+				Log.e("DownloadFileAsync", "doInBackground() exception", e);
+				errorCode = R.string.import_err_cannot_download;
+			}
+			return errorCode;
 		}
 
 		@Override
-		protected void onPostExecute(String result) {
+		protected void onPostExecute(Integer result) {
 			super.onPostExecute(result);
 
-			if (activity.get() == null) {
-				Log.w("DownloadFileAsync", "onPostExecute() skipped -- no activity");
+			if (result == null) { // no error
+				if (activity.get() == null) {
+					Log.w("DownloadFileAsync", "onPostExecute() skipped -- no activity");
+				} else {
+					activity.get().changeStatus(ImportDataActivity.Status.STATUS_DOWNLOAD_COMPLETED);
+				}
+				unzipTask.execute(zipFile, rootDir);
 			} else {
-				activity.get().changeStatus(ImportDataActivity.Status.STATUS_DOWNLOAD_COMPLETED);
+				if (activity.get() != null) {
+					activity.get().displayError(result);
+				}
 			}
-			unzipTask.execute(zipFile, rootDir);
-		}
-
-		@Override
-		protected void onCancelled() {
-			super.onCancelled();
-			// TODO
 		}
 
 		@Override
@@ -94,38 +107,58 @@ public final class ImportDataService {
 			}
 		}
 
-		private void checkMd5() {
+		private Integer checkMd5() {
+			Integer errorCode = null;
+			File md5File = null;
+
 			try {
 				String md5 = Md5.getMd5Sum(zipFile);
-				File md5File = new File(rootDir, TEMP_MD5_FILE);
+				md5File = new File(rootDir, TEMP_MD5_FILE);
 				HttpDownloader downloader = new HttpDownloader(MD5_URL, md5File);
 				downloader.start();
 
 				String remoteMd5 = FileReader.readFile(md5File);
 				if (!md5.equalsIgnoreCase(remoteMd5)) {
-					// TODO
+					Log.e("DownloadFileAsync", "md5 doesn't match");
+					errorCode = R.string.import_err_wrong_dictionary_file;
 				}
-				md5File.delete();
 			} catch (IOException e) {
-				// TODO
+				Log.e("DownloadFileAsync", "checkMd5() exception", e);
+				errorCode = R.string.import_err_wrong_dictionary_file;
+			} finally {
+				if (md5File != null && md5File.exists()) {
+					md5File.delete();
+				}
 			}
+
+			return errorCode;
 		}
 	}
 
-	private class UnzipAsync extends AsyncTask<File, Integer, String> implements Observer {
+	private class UnzipAsync extends AsyncTask<File, Integer, Integer> implements Observer {
 		private File zipFile = null;
 
 		@Override
-		protected String doInBackground(File... params) {
+		protected Integer doInBackground(File... params) {
+			Integer errorCode = null;
+
 			zipFile = params[0];
 			Unzip unzip = new Unzip(zipFile, params[1]);
 			unzip.addObserver(this);
-			unzip.start();
-			return null;
+			try {
+				unzip.start();
+				if (!new File(params[1], DatabaseHelper.DATABASE_NAME).exists()) {
+					throw new IOException("File cannot be found");
+				}
+			} catch (IOException e) {
+				Log.e("UnzipAsync", "checkMd5() exception", e);
+				errorCode = R.string.import_err_wrong_dictionary_file;
+			}
+			return errorCode;
 		}
 
 		@Override
-		protected void onPostExecute(String result) {
+		protected void onPostExecute(Integer result) {
 			super.onPostExecute(result);
 
 			if (zipFile != null && zipFile.exists()) {
@@ -135,14 +168,12 @@ public final class ImportDataService {
 			if (activity.get() == null) {
 				Log.w("UnzipAsync", "onPostExecute() skipped -- no activity");
 			} else {
-				activity.get().changeStatus(ImportDataActivity.Status.STATUS_INSTALL_COMPLETED);
+				if (result == null) { // no error
+					activity.get().changeStatus(ImportDataActivity.Status.STATUS_INSTALL_COMPLETED);
+				} else {
+					activity.get().displayError(result);
+				}
 			}
-		}
-
-		@Override
-		protected void onCancelled() {
-			super.onCancelled();
-			// TODO
 		}
 
 		@Override
